+++ example/test0.m
@@ -1,3 +1,5 @@
 clear
 trig_cycle_20=[0,pi/6,pi/4,pi/3;pi/2,2*pi/3,3*pi/4,5*pi/6;pi,7*pi/6,5*pi/4,4*pi/3;3*pi/2,5*pi/3,7*pi/4,11*pi/6;2*pi,5*pi/2,3*pi,7*pi/2]
 A = 2
+help sub2ind
+X